+++ src/layers_cl/pooling_layer_cl.h
@@ -37,19 +37,19 @@ public:
 
   private:
     bool fuse_relu;
-    size_t input_height;
-    size_t input_width;
-    size_t input_channels;
-    size_t output_height;
-    size_t output_width;
-    size_t output_channels;
-    size_t pad_height;
-    size_t pad_width;
-    size_t kernel_height;
-    size_t kernel_width;
-    size_t stride_height;
-    size_t stride_width;
-    size_t channel_grp_size;
+    uint32_t input_height;
+    uint32_t input_width;
+    uint32_t input_channels;
+    uint32_t output_height;
+    uint32_t output_width;
+    uint32_t output_channels;
+    uint32_t pad_height;
+    uint32_t pad_width;
+    uint32_t kernel_height;
+    uint32_t kernel_width;
+    uint32_t stride_height;
+    uint32_t stride_width;
+    uint32_t channel_grp_size;
     bool global_pooling;
     PoolingParameter_::PoolMethod method;
 };
